Move contextables to own file, remove common
