+++ core/actions/incremental_table.ts
@@ -3,7 +3,6 @@ import {
   ActionBuilder,
   ILegacyBigQueryOptions,
   ILegacyTableConfig,
-  ITableContext,
   LegacyConfigConverter,
   TableType
 } from "df/core/actions";
@@ -28,6 +27,7 @@ import {
   validateQueryString
 } from "df/core/utils";
 import { dataform } from "df/protos/ts";
+import { ITableContext } from "df/core/contextables";
 
 /**
  * When you define an incremental table, Dataform builds the incremental table from scratch only for
